Made more fixes with respect to input format for ML.

+++ cls_builder/cls_builder.h
@@ -17,7 +17,7 @@
  */
 struct Connection {
   tbb::concurrent_unordered_map<std::string, unsigned int> syscall_list_count;
-  tbb::concurrent_unordered_map<std::string, unsigned int> syscall_list_time;
+  tbb::concurrent_unordered_map<std::string, long long> syscall_list_time;
   unsigned int port = -1, tid = -1, pid = -1;
   long long prev = 0;
   std::string ip_addr;
